mon: only share monmap after authenticated
It is no longer necessary to fetch a monmap pre-authentication, something
we previous did for get_monmap_privately().  New code has replaced this
with get_monmap_and_config(), and it authenticates in order to get that
same information (plus configs).

That change was made in mimic, but we must support upgrades from N-2,
which means that luminous daemons still need to function.  The only caller
for get_monmap_privately() in luminous is from ceph-osd during mkfs.
Disabling this here means that new OSDs cannot be created using nautilus
mons and a luminous ceph-osd.  Include a note for the (future) nautilus
upgrade notes.

Reported-by: Christopher Ryan Harrell
Signed-off-by: Sage Weil <sage@redhat.com>