mon/OSDMonitor: don't change in-memory state on prune
We were modifying the in-memory state when running through `do_prune()`.
Should the transaction we're encoding not be committed, we may end up in
a state where we have an in-memory state but did not expect one.

By making sure we work on a temporary state, only to be decoded during
update_from_paxos(), we prevent this sort of inconsistency.

Fixes: http://tracker.ceph.com/issues/24612

Signed-off-by: Joao Eduardo Luis <joao@suse.de>