osd/OSD: choose a fixed thread do oncommits callback function
Now bluestore oncommit callback exec by osd op threads.
If there are multi threads of shard, it will cause out-of order.
For example, threads_per_shard=2
      Thread1                                 Thread2
 swap_oncommits(op1_oncommit)
                                    swap_oncommits(op2_oncommit)
 OpQueueItem.run(Op3)
                                    op2_oncommit.complete();
 op1_oncommit.complete()

This make oncommits out of order.
To avoiding this, we choose a fixed thread which has the smallest
thread_index of shard to do oncommit callback function.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>