mgr/dashboard: Audit REST API calls
Fixes: https://tracker.ceph.com/issues/36193

Enable API auditing with 'ceph dashboard set-audit-api-enabled true' (default is false). If you do not want to log the request payload, then disable it via 'set-audit-api-log-payload false' (default is true).

Example output:
2018-10-08 10:25:21.850994 mgr.x [INF] [DASHBOARD] from='https://[::1]:44410' path='/api/auth' method='POST' user='None' params='{"username": "admin", "password": "admin", "stay_signed_in": false}'

Signed-off-by: Volker Theile <vtheile@suse.com>

Relocate audit logging to tools.RequestLoggingTool().

Signed-off-by: Volker Theile <vtheile@suse.com>

Add documentation.

Signed-off-by: Volker Theile <vtheile@suse.com>

Add JSON processor and let CherryPy decode the request body. This is necessary because the current code tries to read the body multiple times which does not work. With the build-in processor the decoded body is cached an can be processed multiple times.

Signed-off-by: Volker Theile <vtheile@suse.com>